[#538] - Mueve función mapMediaSources y relacionadas a archivo media-sources.functions.ts
