+++ src/api/_utils/functions.ts
@@ -12,16 +12,6 @@ import { baseLanguage } from '../../../cms/utils/localization';
 // Modelos
 import { AuthorDTO } from '@models/author.model';
 import { PrologueDTO } from '@models/prologue.model';
-import { getTweetData } from './twitter-api';
-import {
-	AudioRecording,
-	AudioRecordingSchemaObject,
-	Media,
-	MediaSchemaObject,
-	SpaceRecording,
-	SpaceRecordingSchemaObject,
-	YoutubeVideoSchemaObject,
-} from '@models/media.model';
 
 export function mapAuthor(rawAuthorData: any, language?: string): AuthorDTO {
 	return {
@@ -63,45 +53,4 @@ export function mapResources(resources: any[]) {
 			},
 		})) ?? []
 	);
-}
-
-export async function mapMediaSources(mediaSources: MediaSchemaObject[]): Promise<Media[]> {
-	if (!mediaSources) return [];
-
-	const media: Media[] = [];
-	for (const mediaSource of mediaSources) {
-		if (mediaSource._type === 'audioRecording') {
-			media.push(getAudioRecordingData(mediaSource as AudioRecordingSchemaObject));
-		}
-		if (mediaSource._type === 'spaceRecording') {
-			media.push(await getTweetData(mediaSource as SpaceRecordingSchemaObject));
-		}
-		if (mediaSource._type === 'youtubeRecording') {
-			media.push(getYoutubeVideoData(mediaSource as YoutubeVideoSchemaObject));
-		}
-	}
-	return media;
-}
-
-export function getAudioRecordingData(mediaSource: AudioRecordingSchemaObject): AudioRecording {
-	return {
-		title: mediaSource.title,
-		type: mediaSource._type,
-		icon: mediaSource.icon,
-		data: {
-			url: mediaSource.url,
-		},
-	};
-}
-
-export function getYoutubeVideoData(mediaSource: YoutubeVideoSchemaObject) {
-	return {
-		title: mediaSource.title,
-		type: mediaSource._type,
-		icon: mediaSource.icon,
-		data: {
-			description: mediaSource.description,
-			url: mediaSource.url,
-		},
-	};
 }
